move: 패키지 아키텍쳐 변경
