+++ src/main/java/com/dnd/namuiwiki/domain/question/QuestionRepository.java
@@ -1,6 +1,6 @@
-package com.dnd.namuiwiki.insfrastructure.persistence.question;
+package com.dnd.namuiwiki.domain.question;
 
-import com.dnd.namuiwiki.domain.entity.Question;
+import com.dnd.namuiwiki.domain.question.entity.Question;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface QuestionRepository extends MongoRepository<Question, String> {
